Compatibility with the 3.0.0 specification
The previous version used the slightly newer XSD files, but since
most implementations will be using the published 3.0.0 spec,
shapeshifter-uftp now adheres to this spec as well. In a future version
the updated schema will be supported as well, once it is published by
the shapeshifter organization as a new version.